backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
google-generativeai==0.8.5
